lots of changes
1. Please cut the orange and green

2. Telescoping text is a neat idea but right now feels confusing and
makes some of the bio seem misleading. Please cut for now. Tom says:

"Why not just 'Tom Healy is a writer and poet and …
and then you have to click on something to go further - but you go to
the whole bio all at once.
AND  I think this should be under the house poems, essays whereabouts'"

2.  Don't have the text centered---"Either just left-justified or a mix
of left and right, but not centered pls"

3.  Where "Poems" is followed by the two book covers, change the word
"Poems" to "Poetry"

4.  Change "Where is Tom?" to "Where(About)s"---"Where somehow "about"
is embedded within whereabouts, so you can click on whereabouts to see
my schedule and about to see my bio"

5. Fix Tumblr link. It should link to http://thomashealy.tumblr.com/,
not http://tomhealy.tumblr.com/

6.  And about the contact email---"pls have the email be what my old
website was - I think it was contact@heyday.us or something so that we
could block spam more easily.